requirements.txt
django
openai
python-dotenv
fastapi
uvicorn
requests